% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kerberos.R
\name{check_kerberos}
\alias{check_kerberos}
\title{Kerberos check}
\usage{
check_kerberos()
}
\description{
\code{check_kerberos} check to see if the Kerberos ticket can
be refreshed. If not, it will throw an error.
}

