% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hive-query.R
\name{hive_connect}
\alias{hive_connect}
\title{Connect to a Hive database}
\usage{
hive_connect(db, identifier_quote = "`",
  hive_driver = "org.apache.hive.jdbc.HiveDriver", parameters = "-Xmx2g",
  hadoop_env = list(), default_pars = list(hive.vectorized.execution.enabled
  = "true", hive.vectorized.execution.reduce.enabled = "true",
  hive.resultset.use.unique.column.names = "false", hive.tez.container.size =
  "5120", hive.tez.java.opts = "-Xmx4096m"))
}
\arguments{
\item{db}{Name of the Hive database to connect to}

\item{identifier_quote}{Identifier quote character}

\item{hive_driver}{Hive driver class}

\item{parameters}{String of parameters to use in initializing the JVM}

\item{hadoop_env}{List of Hive parameters to set}

\item{default_pars}{A list of Hive parameters being set by default}
}
\value{
A JDBCConnection object to the Hive database
}
\description{
\code{hive_connect} opens a connection to a Hive database
}

