% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hive-query.R
\name{hive_query}
\alias{hive_query}
\title{Execute a Hive query}
\usage{
hive_query(hive_con, query, clean_names = TRUE, batch = 1e+05,
  update_every = 1, max_rows = Inf, quiet = FALSE)
}
\arguments{
\item{hive_con}{A \code{JDBCConnection} object created by \code{hive_connect}}

\item{query}{A string of the SQL query to run}

\item{clean_names}{Whether to remove database and table name from column names}

\item{batch}{Number of rows to fetch in each batch}

\item{update_every}{How often to print progress updates (after what number of batches)}

\item{max_rows}{The maximum number of rows to fetch from the result}

\item{quiet}{Whether or not to print progress updates}
}
\description{
\code{hive_query} executes a SQL query on a Hive database and pulls down the result
to a local data frame.
}

