% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi-hive.R
\docType{class}
\name{HiveConnection-class}
\alias{HiveConnection-class}
\alias{dbDataType,HiveConnection-method}
\alias{dbSendQuery,HiveConnection,character-method}
\title{HiveConnection class}
\usage{
\S4method{dbSendQuery}{HiveConnection,character}(conn, statement, ...,
  list = NULL)

\S4method{dbDataType}{HiveConnection}(dbObj, obj, ...)
}
\description{
This is a Hive connection class that inherits from 
\code{\link[RJDBC]{JDBCConnection-class}}, which
inherits from \code{\link[DBI]{DBIConnection-class}}
}
\details{
There are custom \code{HiveConnection} methods for \code{dbSendQuery}
and \code{dbDataType}
}

