% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi-hive.R
\docType{class}
\name{HiveResult-class}
\alias{HiveResult-class}
\alias{dbFetch,HiveResult-method}
\title{HiveResult class}
\usage{
\S4method{dbFetch}{HiveResult}(res, n = -1, ...)
}
\description{
This is a Hive result class that inherits from
\code{\link[RJDBC]{JDBCResult-class}}, which inherits
from \code{\link[DBI]{DBIResult-class}}
}
\details{
There is a custom \code{HiveResult} method for \code{dbFetch}
}

