% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle.R
\name{managed_object_types}
\alias{managed_object_types}
\title{Object types whose tables Orpheus generates}
\usage{
managed_object_types(bundle)
}
\arguments{
\item{bundle}{A bundle.}
}
\value{
List of object types.
}
\description{
Object types whose tables Orpheus generates
}
\keyword{internal}
