% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocr.R
\docType{data}
\name{OCR_CHAR_THRESHOLD}
\alias{OCR_CHAR_THRESHOLD}
\title{A page below this many characters is treated as image-only and sent to OCR}
\format{
An object of class \code{integer} of length 1.
}
\usage{
OCR_CHAR_THRESHOLD
}
\description{
A page below this many characters is treated as image-only and sent to OCR
}
\keyword{internal}
