% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\docType{data}
\name{ORPH_ACTIONS}
\alias{ORPH_ACTIONS}
\title{Actions permissions are checked against}
\format{
An object of class \code{character} of length 4.
}
\usage{
ORPH_ACTIONS
}
\description{
Actions permissions are checked against
}
\keyword{datasets}
