% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle.R
\name{orph_active_bundle}
\alias{orph_active_bundle}
\title{Read the active bundle}
\usage{
orph_active_bundle(con)
}
\arguments{
\item{con}{A connection.}
}
\value{
The active bundle list, or \code{NULL} if none is active.
}
\description{
Read the active bundle
}
