% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle.R
\name{orph_default_bundle_path}
\alias{orph_default_bundle_path}
\title{Path to the bundle shipped with the package}
\usage{
orph_default_bundle_path()
}
\description{
Path to the bundle shipped with the package
}
