% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{orph_disconnect}
\alias{orph_disconnect}
\title{Close a connection, releasing the writer lock if held}
\usage{
orph_disconnect(con)
}
\arguments{
\item{con}{A connection from \code{\link[=orph_connect]{orph_connect()}}.}
}
\description{
Close a connection, releasing the writer lock if held
}
