% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\docType{data}
\name{ORPH_DOC_TYPES}
\alias{ORPH_DOC_TYPES}
\title{Document types the classifier chooses between}
\format{
An object of class \code{character} of length 5.
}
\usage{
ORPH_DOC_TYPES
}
\description{
Document types the classifier chooses between
}
\keyword{datasets}
