% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{ORPH_EXCLUDED_STATUSES}
\alias{ORPH_EXCLUDED_STATUSES}
\title{Statuses whose rows are excluded from downstream use}
\format{
An object of class \code{character} of length 1.
}
\usage{
ORPH_EXCLUDED_STATUSES
}
\description{
Statuses whose rows are excluded from downstream use
}
\keyword{internal}
