% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{orph_get_actor}
\alias{orph_get_actor}
\title{Read an actor}
\usage{
orph_get_actor(con, actor_id)
}
\arguments{
\item{con}{A connection.}

\item{actor_id}{Actor identifier.}
}
\value{
A list, or \code{NULL}.
}
\description{
Read an actor
}
