% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{orph_get_document}
\alias{orph_get_document}
\title{Document metadata}
\usage{
orph_get_document(con, document_id)
}
\arguments{
\item{con}{A connection.}

\item{document_id}{Document identifier.}
}
\value{
A one-row list, or \code{NULL}.
}
\description{
Document metadata
}
