% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle.R
\name{orph_implementing_types}
\alias{orph_implementing_types}
\title{Object types that implement an interface}
\usage{
orph_implementing_types(bundle, interface_id)
}
\arguments{
\item{bundle}{A bundle.}

\item{interface_id}{Interface identifier.}
}
\value{
Character vector of object type ids.
}
\description{
Object types that implement an interface
}
