% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle.R
\name{orph_interface}
\alias{orph_interface}
\title{Look up an interface in a bundle}
\usage{
orph_interface(bundle, interface_id)
}
\arguments{
\item{bundle}{A bundle.}

\item{interface_id}{Interface identifier.}
}
\value{
The interface list, or \code{NULL}.
}
\description{
Look up an interface in a bundle
}
