% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle.R
\name{orph_interface_property_ids}
\alias{orph_interface_property_ids}
\title{Property identifiers an interface requires}
\usage{
orph_interface_property_ids(interface)
}
\arguments{
\item{interface}{An interface list.}
}
\value{
Character vector.
}
\description{
Property identifiers an interface requires
}
