% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle.R
\name{orph_link_type}
\alias{orph_link_type}
\title{Look up a link type in a bundle}
\usage{
orph_link_type(bundle, link_type_id)
}
\arguments{
\item{bundle}{A bundle.}

\item{link_type_id}{Link type identifier.}
}
\value{
The link type list, or \code{NULL}.
}
\description{
Look up a link type in a bundle
}
