% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{orph_list_documents}
\alias{orph_list_documents}
\title{List documents}
\usage{
orph_list_documents(con, limit = 100L)
}
\arguments{
\item{con}{A connection.}

\item{limit}{Maximum rows.}
}
\value{
A data frame.
}
\description{
List documents
}
