% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{orph_migrate}
\alias{orph_migrate}
\title{Apply pending migrations}
\usage{
orph_migrate(con)
}
\arguments{
\item{con}{A writable connection.}
}
\value{
Invisibly, the versions applied.
}
\description{
Apply pending migrations
}
