% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\docType{data}
\name{ORPH_NAIVE_RESOLUTION}
\alias{ORPH_NAIVE_RESOLUTION}
\title{Quality label attached to every naive-matched result}
\format{
An object of class \code{character} of length 1.
}
\usage{
ORPH_NAIVE_RESOLUTION
}
\description{
Quality label attached to every naive-matched result
}
\keyword{datasets}
