% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle.R
\name{orph_object_type}
\alias{orph_object_type}
\title{Look up an object type in a bundle}
\usage{
orph_object_type(bundle, type_id)
}
\arguments{
\item{bundle}{A bundle.}

\item{type_id}{Object type identifier.}
}
\value{
The object type list, or \code{NULL}.
}
\description{
Look up an object type in a bundle
}
