% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle.R
\name{orph_property_ids}
\alias{orph_property_ids}
\title{Property identifiers of an object type}
\usage{
orph_property_ids(object_type)
}
\arguments{
\item{object_type}{An object type list.}
}
\value{
Character vector.
}
\description{
Property identifiers of an object type
}
