% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amend.R
\name{orph_review_progress}
\alias{orph_review_progress}
\title{Per-document review progress}
\usage{
orph_review_progress(con, document_id)
}
\arguments{
\item{con}{A connection.}

\item{document_id}{Document identifier.}
}
\value{
A list of counts by status.
}
\description{
Per-document review progress
}
