% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{orph_set_setting}
\alias{orph_set_setting}
\title{Write an org-level setting}
\usage{
orph_set_setting(con, key, value, actor_id = NULL)
}
\arguments{
\item{con}{A writable connection.}

\item{key}{Setting key.}

\item{value}{Setting value.}

\item{actor_id}{Actor making the change.}
}
\description{
Write an org-level setting
}
