% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\docType{data}
\name{ORPH_SHARE_ROLES}
\alias{ORPH_SHARE_ROLES}
\title{Roles a document share can grant}
\format{
An object of class \code{character} of length 2.
}
\usage{
ORPH_SHARE_ROLES
}
\description{
Roles a document share can grant
}
\keyword{datasets}
