% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{ORPH_SOURCES}
\alias{ORPH_SOURCES}
\title{Provenance source values}
\format{
An object of class \code{character} of length 3.
}
\usage{
ORPH_SOURCES
}
\description{
Provenance source values
}
\keyword{datasets}
