% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{ORPH_STATUSES}
\alias{ORPH_STATUSES}
\title{Instance review status values}
\format{
An object of class \code{character} of length 4.
}
\usage{
ORPH_STATUSES
}
\description{
Instance review status values
}
\keyword{datasets}
