% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\docType{data}
\name{ORPH_VISIBILITY}
\alias{ORPH_VISIBILITY}
\title{Document visibility levels}
\format{
An object of class \code{character} of length 3.
}
\usage{
ORPH_VISIBILITY
}
\description{
Document visibility levels
}
\keyword{datasets}
