% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasette.R
\name{orph_write_datasette_metadata}
\alias{orph_write_datasette_metadata}
\title{Write a Datasette metadata file for the store}
\usage{
orph_write_datasette_metadata(
  path = "inst/datasette/metadata.yml",
  database_name = "orpheus"
)
}
\arguments{
\item{path}{Where to write the YAML.}

\item{database_name}{The name Datasette will serve the database under.}
}
\value{
Invisibly, \code{path}.
}
\description{
Write a Datasette metadata file for the store
}
