% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-write-hdfs.R
\name{read_hdfs}
\alias{read_hdfs}
\title{Read a file on HDFS into a data frame.}
\usage{
read_hdfs(path, read_fn = readr::read_delim, ..., file_argnum = 1)
}
\arguments{
\item{path}{File path on HDFS of the file to read in.}

\item{read_fn}{Function to use to read the file (\code{readr::read_delim} by default)}

\item{...}{Additional parameters to pass to \code{read_fn}}

\item{file_argnum}{The position of the file path argument in \code{read_fn}
(e.g., \code{file} is the first argument in \code{readr::read_delim},
\code{path} is the first argument in \code{readxl::read_excel})}
}
\value{
The object read by \code{read_fn}.
}
\description{
\code{read_hdfs} reads a file on HDFS into a data frame
}
\details{
This function reads a file on HDFS into a data frame. By default, it will use
\code{read_delim} in \code{readr} to read the file. If no delimiter is specified
and \code{read_delim} is used, it will use a comma as the delimiter.
}

