% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concepts.R
\name{register_concept_templates}
\alias{register_concept_templates}
\title{Register the bundle's concept templates with conceptR}
\usage{
register_concept_templates(con, bundle, ctx)
}
\description{
Register the bundle's concept templates with conceptR
}
\keyword{internal}
