% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\docType{data}
\name{REVIEWED_STATUSES}
\alias{REVIEWED_STATUSES}
\title{Statuses that count as a human having judged a row}
\format{
An object of class \code{character} of length 3.
}
\usage{
REVIEWED_STATUSES
}
\description{
Statuses that count as a human having judged a row
}
\keyword{internal}
