% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{with_tx}
\alias{with_tx}
\title{Run a function inside a transaction}
\usage{
with_tx(con, expr)
}
\arguments{
\item{con}{A writable connection.}

\item{expr}{Expression to evaluate.}
}
\description{
Run a function inside a transaction
}
\keyword{internal}
