% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-write-hdfs.R
\name{write_hdfs}
\alias{write_hdfs}
\title{Write a data frame to a file on HDFS.}
\usage{
write_hdfs(df, path, write_fn = readr::write_delim, ..., df_argnum = 1,
  file_argnum = 2)
}
\arguments{
\item{df}{A data frame.}

\item{path}{File path on HDFS to write to}

\item{write_fn}{Function to use to write the data frame to a file (\code{readr::write_csv} by default)}

\item{...}{Additional paramters to pass to \code{write_csv}}

\item{df_argnum}{The position of the data frame argument in \code{write_fn}
(e.g., \code{x} is the first argument in \code{readr::write_csv},
\code{object} is the first argument in \code{saveRDS})}

\item{file_argnum}{The position of the file path argument in \code{write_fn}
(e.g., \code{path} is the second argument in \code{readr::write_csv}),
\code{file} is the second argument in \code{saveRDS}}
}
\description{
Write a data frame to a file on HDFS.
}

