% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{writer_lock_path}
\alias{writer_lock_path}
\title{Path of the advisory writer lock for a database}
\usage{
writer_lock_path(path)
}
\description{
Path of the advisory writer lock for a database
}
\keyword{internal}
